refactor: separate @vrew/modules/web-bridge/types.ts
