+++ packages/modules/web-bridge/types/action.ts
@@ -1,0 +1,14 @@
+export enum BridgeActions {
+  NAVIGATION_NAVIGATE = "navigation-navigate",
+  NAVIGATION_GO_BACK = "navigation-go-back",
+  NAVIGATION_CAN_GO_BACK = "navigation-can-go-back",
+  NAVIGATION_PUSH = "navigation-push",
+  NAVIGATION_POP_TO_TOP = "navigation-pop-to-top",
+  NAVIGATION_POP = "navigation-pop",
+  NAVIGATION_REPLACE = "navigation-replace",
+  NAVIGATION_RESET = "navigation-reset",
+  NAVIGATION_RELOAD = "navigation-reload",
+  NAVIGATION_SET_OPTIONS = "navigation-set-options",
+
+  CONSOLE_LOG = "console-log",
+}
